남후승

+++ KOSTA_3_final/src/main/java/com/kosta/KOSTA_3_final/persistance/product_package/PackageRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.kosta.KOSTA_3_final.model.product_package.PackageVO;
 
-public interface PackageRepository extends CrudRepository<PackageVO, Integer>{
+public interface PackageRepository extends CrudRepository<PackageVO, Long>{
 	
 }
